+++ components/infrared/impl/src/main/kotlin/com/flipperdevices/infrared/impl/model/InfraredNavigationConfig.kt
@@ -13,4 +13,7 @@ sealed class InfraredNavigationConfig {
 
     @Serializable
     data class Rename(val keyPath: FlipperKeyPath) : InfraredNavigationConfig()
+
+    @Serializable
+    data class RemoteControl(val keyPath: FlipperKeyPath) : InfraredNavigationConfig()
 }