+++ components/keyedit/noop/src/main/java/com/flipperdevices/keyedit/KeyEditDecomposeComponentNoop.kt
@@ -2,6 +2,7 @@ package com.flipperdevices.keyedit
 
 import androidx.compose.runtime.Composable
 import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.bridge.dao.api.model.FlipperKey
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.keyedit.api.KeyEditDecomposeComponent
@@ -23,6 +24,7 @@ class KeyEditDecomposeComponentNoop(
         override fun invoke(
             componentContext: ComponentContext,
             onBack: DecomposeOnBackParameter,
+            onSave: (FlipperKey?) -> Unit,
             flipperKeyPath: FlipperKeyPath,
             title: String?
         ) = KeyEditDecomposeComponentNoop(componentContext)
@@ -30,6 +32,7 @@ class KeyEditDecomposeComponentNoop(
         override fun invoke(
             componentContext: ComponentContext,
             onBack: DecomposeOnBackParameter,
+            onSave: (FlipperKey?) -> Unit,
             notSavedFlipperKey: NotSavedFlipperKey,
             title: String?
         ) = KeyEditDecomposeComponentNoop(componentContext)