+++ components/remote-controls/brands/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/brands/composable/BrandsScreen.kt
@@ -11,9 +11,9 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import com.flipperdevices.ifrmvp.core.ui.layout.shared.ErrorComposable
-import com.flipperdevices.ifrmvp.core.ui.layout.shared.LoadingComposable
 import com.flipperdevices.ifrmvp.core.ui.layout.shared.SharedTopBar
 import com.flipperdevices.remotecontrols.impl.brands.composable.composable.BrandsLoadedContent
+import com.flipperdevices.remotecontrols.impl.brands.composable.composable.BrandsLoadingComposable
 import com.flipperdevices.remotecontrols.impl.brands.presentation.decompose.BrandsDecomposeComponent
 import com.flipperdevices.remotecontrols.brands.impl.R as BrandsR
 
@@ -51,7 +51,7 @@ fun BrandsScreen(
                 }
 
                 BrandsDecomposeComponent.Model.Loading -> {
-                    LoadingComposable()
+                    BrandsLoadingComposable()
                 }
             }
         }