+++ components/remote-controls/core-ui/src/main/kotlin/com/flipperdevices/ifrmvp/core/ui/button/Base64ImageButton.kt
@@ -64,7 +64,6 @@ fun rememberImageBitmap(base64Image: String): ImageBitmap? {
 @Composable
 fun Base64ImageButton(
     base64Icon: String,
-    isEmulating: Boolean,
     modifier: Modifier = Modifier,
     onClick: () -> Unit
 ) {
@@ -76,9 +75,10 @@ fun Base64ImageButton(
             bitmap = imageBitmap,
             iconTint = Color.Unspecified,
             modifier = modifier,
-            isEmulating = isEmulating
         )
     } else {
-        UnknownButton(onClick = onClick)
+        UnknownButton(
+            onClick = onClick,
+        )
     }
 }