+++ components/remote-controls/core-ui/src/main/kotlin/com/flipperdevices/ifrmvp/core/ui/button/DoubleButton.kt
@@ -9,6 +9,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import com.flipperdevices.core.ui.theme.LocalPalletV2
+import com.flipperdevices.ifrmvp.core.ui.button.core.LocalButtonPlaceholder
+import com.flipperdevices.ifrmvp.core.ui.button.core.SyncingBox
 import com.flipperdevices.ifrmvp.core.ui.button.core.TextButton
 import com.flipperdevices.ifrmvp.core.ui.layout.core.sf
 
@@ -18,7 +20,6 @@ fun DoubleButton(
     onLastClick: () -> Unit,
     firstText: String,
     lastText: String,
-    isEmulating: Boolean,
     modifier: Modifier = Modifier,
     text: String? = null,
 ) {
@@ -33,22 +34,20 @@ fun DoubleButton(
             onClick = onFirstClick,
             text = firstText,
             background = LocalPalletV2.current.surface.menu.body.dufault,
-            isEmulating = isEmulating,
         )
         text?.let {
             TextButton(
                 onClick = null,
                 text = text,
                 background = LocalPalletV2.current.surface.menu.body.dufault,
-                isEmulating = isEmulating
             )
         }
         TextButton(
             onClick = onLastClick,
             text = lastText,
             background = LocalPalletV2.current.surface.menu.body.dufault,
-            isEmulating = isEmulating
         )
+        SyncingBox(LocalButtonPlaceholder.current.isSyncing)
     }
 }
 
@@ -56,7 +55,6 @@ fun DoubleButton(
 fun VolumeButton(
     onAddClick: () -> Unit,
     onReduceClick: () -> Unit,
-    isEmulating: Boolean,
     modifier: Modifier = Modifier
 ) {
     DoubleButton(
@@ -66,7 +64,6 @@ fun VolumeButton(
         firstText = "+",
         lastText = "-",
         modifier = modifier,
-        isEmulating = isEmulating
     )
 }
 
@@ -74,7 +71,6 @@ fun VolumeButton(
 fun ChannelButton(
     onNextClick: () -> Unit,
     onPrevClick: () -> Unit,
-    isEmulating: Boolean,
     modifier: Modifier = Modifier,
 ) {
     DoubleButton(
@@ -84,6 +80,5 @@ fun ChannelButton(
         firstText = "+",
         lastText = "-",
         modifier = modifier,
-        isEmulating = isEmulating
     )
 }