+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/presentation/viewmodel/SaveTempSignalViewModel.kt
@@ -1,8 +1,6 @@
 package com.flipperdevices.remotecontrols.impl.setup.presentation.viewmodel
 
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
-import com.flipperdevices.bridge.service.api.FlipperServiceApi
-import com.flipperdevices.bridge.service.api.provider.FlipperBleServiceConsumer
 import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ktx.jre.FlipperDispatchers
@@ -30,7 +28,6 @@ class SaveTempSignalViewModel @Inject constructor(
     private val saveFileApi: SaveFileApi,
     private val saveFolderApi: SaveFolderApi,
 ) : DecomposeViewModel(),
-    FlipperBleServiceConsumer,
     LogTagProvider,
     SaveTempSignalApi {
     private val mutex = Mutex()
@@ -38,50 +35,42 @@ class SaveTempSignalViewModel @Inject constructor(
     private val _state = MutableStateFlow<SaveTempSignalApi.State>(SaveTempSignalApi.State.Pending)
     override val state = _state.asStateFlow()
 
-    override fun saveFile(
-        textContent: String,
-        nameWithExtension: String,
-        extFolderPath: String
-    ) = save(
-        extFolderPath = extFolderPath,
-        textContent = textContent,
-        absolutePath = FlipperFilePath(
-            folder = extFolderPath,
-            nameWithExtension = nameWithExtension
-        ).getPathOnFlipper(),
-    )
-
-    private fun save(
-        textContent: String,
-        absolutePath: String,
-        extFolderPath: String
+    override fun saveFiles(
+        vararg filesDesc: SaveTempSignalApi.FileDesc,
+        onFinished: () -> Unit,
     ) {
         viewModelScope.launch {
             _state.emit(SaveTempSignalApi.State.Uploading(0, 0))
             val serviceApi = serviceProvider.getServiceApi()
             launchWithLock(mutex, viewModelScope, "load") {
-                saveFolderApi.save(serviceApi.requestApi, "$EXT_PATH/$extFolderPath")
-                val saveFileFlow = saveFileApi.save(
-                    requestApi = serviceApi.requestApi,
-                    textContent = textContent,
-                    absolutePath = absolutePath
-                )
-                saveFileFlow
-                    .flowOn(FlipperDispatchers.workStealingDispatcher)
-                    .onEach {
-                        _state.value = when (it) {
-                            SaveFileApi.Status.Finished -> SaveTempSignalApi.State.Uploaded
-                            is SaveFileApi.Status.Saving -> SaveTempSignalApi.State.Uploading(
-                                it.uploaded,
-                                it.size
-                            )
+                filesDesc.forEach { fileDesc ->
+                    val absolutePath = FlipperFilePath(
+                        folder = fileDesc.extFolderPath,
+                        nameWithExtension = fileDesc.nameWithExtension
+                    ).getPathOnFlipper()
+
+                    saveFolderApi.save(serviceApi.requestApi, "$EXT_PATH/${fileDesc.extFolderPath}")
+                    val saveFileFlow = saveFileApi.save(
+                        requestApi = serviceApi.requestApi,
+                        textContent = fileDesc.textContent,
+                        absolutePath = absolutePath
+                    )
+                    saveFileFlow
+                        .flowOn(FlipperDispatchers.workStealingDispatcher)
+                        .onEach {
+                            _state.value = when (it) {
+                                SaveFileApi.Status.Finished -> SaveTempSignalApi.State.Uploaded
+                                is SaveFileApi.Status.Saving -> SaveTempSignalApi.State.Uploading(
+                                    it.uploaded,
+                                    it.size
+                                )
+                            }
                         }
-                    }
-                    .collect()
+                        .collect()
+                }
                 _state.value = SaveTempSignalApi.State.Uploaded
+                onFinished.invoke()
             }
         }
     }
-
-    override fun onServiceApiReady(serviceApi: FlipperServiceApi) = Unit
 }